Remove camel case mutation on override names
This introduced surprising behaviour whereby the actual overridden name
got mutated in a way the end user might not expected. For example,
setting a name to FixedAPEnabled would end up re-writing it to
FixedApenabled, which is not the correct outcome.

Instead it's now expected that the user is going to do the right thing